pre/pyproject.toml
[project]
name = "pre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "facenet-pytorch>=2.5.3",
    "flwr>=1.27.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=9.5.0",
    "streamlit>=1.55.0",
    "streamlit-webrtc>=0.64.5",
    "tabulate>=0.10.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.67.3",
]
